[Chore #163083232] Change directory name for table schema
