+++ server/middlewares/schema/schemas/index.js
@@ -1,0 +1,10 @@
+import meetupSchemaValidator from './meetup';
+import questionSchemaValidator from './question';
+import rsvpSchemaValidator from './rsvp';
+
+export default {
+  '/meetups': meetupSchemaValidator,
+  '/questions': questionSchemaValidator,
+  '/meetups/:meetupId/rsvps': rsvpSchemaValidator,
+  '/meetups/:meetupId/rsvps/:rsvpId': rsvpSchemaValidator
+};
