+++ server/middlewares/schema/schemas/question.js
@@ -1,0 +1,10 @@
+const Joi = require('joi');
+
+const questionDataSchema = Joi.object().keys({
+  title: Joi.string().strict().required(),
+  body: Joi.string().required(),
+  meetupId: Joi.number().integer().required(),
+  userId: Joi.number().integer().required()
+});
+
+export default questionDataSchema;
