+++ server/middlewares/schema/schemas/rsvp.js
@@ -1,0 +1,8 @@
+const Joi = require('joi');
+
+const rsvpDataSchema = Joi.object().keys({
+  userId: Joi.number().integer().required(),
+  response: Joi.string().allow(['yes', 'no', 'maybe']).strict().required(),
+});
+
+export default rsvpDataSchema;
