--- server/models/helpers/meetup.js
@@ -1,13 +1,0 @@
-export default {
-  text: `CREATE TABLE IF NOT EXISTS Meetup (
-      id SERIAL PRIMARY KEY,
-      topic text NOT NULL,
-      location text NOT NULL,
-      happeningOn DATE NOT NULL,
-      CHECK (happeningOn >= NOW()),
-      createdOn DATE DEFAULT NOW(),
-      images text[],
-      tags text[],
-      maxNumberOfAttendees INTEGER DEFAULT 50
-    )`
-};
