+++ server/models/schemas/index.js
@@ -1,10 +1,26 @@
-import meetupSchemaValidator from './meetup';
-import questionSchemaValidator from './question';
-import rsvpSchemaValidator from './rsvp';
+/**
+ * @module
+ * @description Helper code specific to models
+ */
 
-export default {
-  '/meetups': meetupSchemaValidator,
-  '/questions': questionSchemaValidator,
-  '/meetups/:meetupId/rsvps': rsvpSchemaValidator,
-  '/meetups/:meetupId/rsvps/:rsvpId': rsvpSchemaValidator
+import createMeetupSQLQuery from './meetup';
+import createQuestionSQLQuery from './question';
+import createUserSQLQuery from './user';
+import createRsvpSQLQuery from './rsvp';
+import createCommentSQLQuery from './comment';
+import createVoteSQLQuery from './vote';
+import createUpvoteSQLQuery from './upvote';
+import createDownvoteSQLQuery from './downvote';
+
+const createTableSQLQueries = {
+  Meetup: createMeetupSQLQuery,
+  Question: createQuestionSQLQuery,
+  User: createUserSQLQuery,
+  Rsvp: createRsvpSQLQuery,
+  Comment: createCommentSQLQuery,
+  Vote: createVoteSQLQuery,
+  Upvote: createUpvoteSQLQuery,
+  Downvote: createDownvoteSQLQuery
 };
+
+export default createTableSQLQueries;
