+++ server/models/schemas/question.js
@@ -1,10 +1,11 @@
-const Joi = require('joi');
-
-const questionDataSchema = Joi.object().keys({
-  title: Joi.string().strict().required(),
-  body: Joi.string().required(),
-  meetupId: Joi.number().integer().required(),
-  userId: Joi.number().integer().required()
-});
-
-export default questionDataSchema;
+export default {
+  text: `CREATE TABLE IF NOT EXISTS Question (
+      id SERIAL PRIMARY KEY,
+      title text NOT NULL,
+      body text NOT NULL,
+      createdBy INTEGER NOT NULL REFERENCES "User",
+      meetup INTEGER NOT NULL REFERENCES Meetup,
+      votes INTEGER DEFAULT 0 CONSTRAINT positive_votes CHECK (votes >= 0),
+      createdOn DATE DEFAULT NOW()
+    )`
+};
