+++ server/models/schemas/rsvp.js
@@ -1,8 +1,8 @@
-const Joi = require('joi');
-
-const rsvpDataSchema = Joi.object().keys({
-  userId: Joi.number().integer().required(),
-  response: Joi.string().allow(['yes', 'no', 'maybe']).strict().required(),
-});
-
-export default rsvpDataSchema;
+export default {
+  text: `CREATE TABLE IF NOT EXISTS Rsvp (
+      id SERIAL NOT NULL PRIMARY KEY,
+      "user" INTEGER NOT NULL REFERENCES "User",
+      meetup INTEGER NOT NULL REFERENCES Meetup,
+      response text NOT NULL
+    )`
+};
