+++ server/models/schemas/upvote.js
