+++ server/models/schemas/vote.js
